Moved getExtFromUrl to path.ts
